requirements.txt
langchain
langchain-community
langchain-groq
langchain-huggingface
chromadb
sentence-transformers
pypdf
sqlalchemy
streamlit
python-dotenv
tiktoken
